% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_expand.R
\name{postal_expand}
\alias{postal_expand}
\title{postal_expand
Converts messy real-world addresses into normalized equivalents suitable for search indexing, hashing, etc.}
\usage{
postal_expand(adr, libpostal_path = "~/libpostal/src/")
}
\arguments{
\item{adr}{string}

\item{libpostal_path}{path to libpostal (default : ~/libpostal/src/)}
}
\value{
Adresses expansions
}
\description{
postal_expand
Converts messy real-world addresses into normalized equivalents suitable for search indexing, hashing, etc.
}
\examples{
adr1 <- "Quatre vignt douze Ave des Champs-Élysées"
adr2 <- "sept rue des freres lumiere, Rueil-Malmaison"

postal_expand(adr1)
#$`Quatre vignt douze Ave des Champs-Élysées.expansions`
#[1] "92 avenue des champs-elysees" 
#[2] "92 avenue des champs elysees"
#[3] "92 avenue des champselysees" 

postal_expand(c(adr1,adr2))
}

