% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal_parse.R
\name{postal_parse}
\alias{postal_parse}
\title{postal_parse
Parse address into components.}
\usage{
postal_parse(adr, libpostal_path = "~/libpostal/src/")
}
\arguments{
\item{adr}{string}

\item{libpostal_path}{path to libpostal (default : ~/libpostal/src/)}
}
\value{
Adresses components
}
\description{
postal_parse
Parse address into components.
}
\examples{
postal_parse(c("Quatre vignt douze Ave des Champs-Élysées","sept rue des freres lumiere, Rueil-Malmaison"))
}

